integration/pyproject.toml
[project]
name = "tradewinds-api-client"
version = "1.0"
description = "A client library for accessing Tradewinds API"
authors = []
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
    "httpx>=0.23.0,<0.29.0",
    "attrs>=22.2.0",
    "python-dateutil>=2.8.0,<3",
]

[tool.uv.build-backend]
module-name = "tradewinds_api_client"
module-root = ""

[build-system]
requires = ["uv_build>=0.10.0,<0.11.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
